llm-translate/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 分词工具
sentencepiece>=0.1.99

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 可视化
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 评估指标
sacrebleu>=2.3.1
nltk>=3.8.1

# 工具库
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# 日志
loguru>=0.7.0

# 部署
onnx>=1.14.0
onnxruntime>=1.15.0

# 可选：分布式训练
accelerate>=0.20.0

# 开发工具
pytest>=7.4.0
black>=23.0.0
